+++ src/Rules.php
@@ -68,6 +68,14 @@ final class Rules
     }
 
     /**
+     * {@inheritdoc}
+     */
+    public static function __set_state(array $properties): self
+    {
+        return new self($properties['rules']);
+    }
+
+    /**
      * new instance.
      *
      * @param array $rules
@@ -78,7 +86,43 @@ final class Rules
     }
 
     /**
-     * Returns PSL ICANN public info for a given domain.
+     * Tell whether the submitted section if supported by the current object
+     *
+     * @param string $section
+     *
+     * @return bool
+     */
+    public function supports(string $section): bool
+    {
+        return self::ALL_DOMAINS === $section
+            || (isset($this->rules[$section]) && is_array($this->rules[$section]) && !empty($this->rules[$section]));
+    }
+
+    /**
+     * Determines the public suffix for a given domain.
+     *
+     * @param string|null $domain
+     * @param string      $section
+     *
+     * @throws Exception
+     *                   If the Domain is invalid or malformed
+     *                   If the section is invalid or not supported
+     *                   If the PublicSuffix can not be converted using against the domain encoding.
+     *
+     * @return PublicSuffix
+     */
+    public function getPublicSuffix(string $domain = null, string $section = self::ALL_DOMAINS): PublicSuffix
+    {
+        if (!$this->isMatchable($domain)) {
+            throw new Exception(sprintf('The submitted domain `%s` is invalid or malformed', $domain));
+        }
+        $this->validateSection($section);
+
+        return $this->findPublicSuffix($domain, $section);
+    }
+
+    /**
+     * Returns PSL info for a given domain.
      *
      * @param string|null $domain
      * @param string      $section
@@ -87,20 +131,16 @@ final class Rules
      */
     public function resolve(string $domain = null, string $section = self::ALL_DOMAINS): Domain
     {
-        if (!in_array($section, [self::PRIVATE_DOMAINS, self::ICANN_DOMAINS, self::ALL_DOMAINS], true)) {
-            throw new Exception(sprintf('%s is an unknown Public Suffix List section', $section));
-        }
-
+        $this->validateSection($section);
         if (!$this->isMatchable($domain)) {
             return new Domain();
         }
 
-        $publicSuffix = $this->findPublicSuffix($domain, $section);
-        if (null === $publicSuffix->getContent()) {
-            return new Domain($domain, $this->handleNoMatches($domain));
+        try {
+            return new Domain($domain, $this->findPublicSuffix($domain, $section));
+        } catch (Exception $e) {
+            return new Domain($domain);
         }
-
-        return new Domain($domain, $this->handleMatches($domain, $publicSuffix));
     }
 
     /**
@@ -119,28 +159,17 @@ final class Rules
     }
 
     /**
-     * Normalizes a domain name.
+     * Assert the section status.
      *
-     * "The domain must be canonicalized in the normal way for hostnames - lower-case, Punycode."
+     * @param string $section
      *
-     * @see https://tools.ietf.org/html/rfc3492
-     *
-     * @param string $domain
-     *
-     * @return string
+     * @throws Exception if the submitted section is not supported
      */
-    private function normalize(string $domain): string
+    private function validateSection(string $section)
     {
-        if (false !== strpos($domain, '%')) {
-            $domain = rawurldecode($domain);
+        if (!$this->supports($section)) {
+            throw new Exception(sprintf('%s is an unknown Public Suffix List section', $section));
         }
-
-        $normalize = idn_to_ascii($domain, 0, INTL_IDNA_VARIANT_UTS46);
-        if (false === $normalize) {
-            return '';
-        }
-
-        return strtolower($normalize);
     }
 
     /**
@@ -153,23 +182,47 @@ final class Rules
      */
     private function findPublicSuffix(string $domain, string $section): PublicSuffix
     {
-        $normalizedDomain = $this->normalize($domain);
-        $reverseLabels = array_reverse(explode('.', $normalizedDomain));
-        $resultIcann = $this->findPublicSuffixFromSection($reverseLabels, self::ICANN_DOMAINS);
+        $reverseLabels = array_reverse(explode('.', $this->normalizeDomain($domain)));
+        $icann = $this->findPublicSuffixFromSection($reverseLabels, self::ICANN_DOMAINS);
         if (self::ICANN_DOMAINS === $section) {
-            return $resultIcann;
+            return $this->normalizePublicSuffix($icann, $domain);
         }
 
-        $resultPrivate = $this->findPublicSuffixFromSection($reverseLabels, self::PRIVATE_DOMAINS);
-        if (count($resultPrivate) > count($resultIcann)) {
-            return $resultPrivate;
+        $private = $this->findPublicSuffixFromSection($reverseLabels, self::PRIVATE_DOMAINS);
+        if (count($private) > count($icann)) {
+            return $this->normalizePublicSuffix($private, $domain);
         }
 
         if (self::ALL_DOMAINS === $section) {
-            return $resultIcann;
+            return $this->normalizePublicSuffix($icann, $domain);
         }
 
-        return new PublicSuffix();
+        return $this->normalizePublicSuffix(new PublicSuffix(), $domain);
+    }
+
+    /**
+     * Normalizes a domain name.
+     *
+     * "The domain must be canonicalized in the normal way for hostnames - lower-case, Punycode."
+     *
+     * @see https://tools.ietf.org/html/rfc3492
+     *
+     * @param string $domain
+     *
+     * @return string
+     */
+    private function normalizeDomain(string $domain): string
+    {
+        if (false !== strpos($domain, '%')) {
+            $domain = rawurldecode($domain);
+        }
+
+        $normalize = idn_to_ascii($domain, 0, INTL_IDNA_VARIANT_UTS46, $arr);
+        if ($arr['errors'] > 0) {
+            return '';
+        }
+
+        return strtolower($normalize);
     }
 
     /**
@@ -182,7 +235,7 @@ final class Rules
      */
     private function findPublicSuffixFromSection(array $labels, string $section): PublicSuffix
     {
-        $rules = $this->rules[$section] ?? null;
+        $rules = $this->rules[$section] ?? [];
         $matches = [];
         foreach ($labels as $label) {
             //match exception rule
@@ -214,57 +267,24 @@ final class Rules
     }
 
     /**
-     * Returns a PublicSuffix if none was found using the PSL.
+     * Normalize the found Public Suffix against its domain name.
      *
-     * @param string $domain
-     *
-     * @return PublicSuffix
-     */
-    private function handleNoMatches(string $domain): PublicSuffix
-    {
-        $labels = explode('.', $domain);
-        $publicSuffix = array_pop($labels);
-        if ($this->isPunycoded($domain)) {
-            return new PublicSuffix($publicSuffix);
-        }
-
-        $publicSuffix = idn_to_utf8($publicSuffix, 0, INTL_IDNA_VARIANT_UTS46);
-        if (false !== $publicSuffix) {
-            return new PublicSuffix($publicSuffix);
-        }
-
-        return new PublicSuffix();
-    }
-
-    /**
-     * Tells whether the domain is punycoded.
-     *
-     * @param string $domain
-     *
-     * @return bool
-     */
-    private function isPunycoded(string $domain): bool
-    {
-        return false !== strpos($domain, 'xn--');
-    }
-
-    /**
-     * Returns a PublicSuffix if one was found using the PSL.
-     *
-     * @param string       $domain
      * @param PublicSuffix $publicSuffix
+     * @param string       $domain
      *
      * @return PublicSuffix
      */
-    private function handleMatches($domain, PublicSuffix $publicSuffix): PublicSuffix
+    private function normalizePublicSuffix(PublicSuffix $publicSuffix, string $domain): PublicSuffix
     {
-        if ($this->isPunycoded($domain)) {
-            return $publicSuffix;
+        if (null === $publicSuffix->getContent()) {
+            $labels = explode('.', $domain);
+            $publicSuffix = new PublicSuffix(array_pop($labels));
         }
 
-        return new PublicSuffix(
-            idn_to_utf8($publicSuffix->getContent(), 0, INTL_IDNA_VARIANT_UTS46),
-            $publicSuffix->isICANN() ? self::ICANN_DOMAINS : self::PRIVATE_DOMAINS
-        );
+        if (false === strpos($domain, 'xn--')) {
+            return $publicSuffix->toUnicode();
+        }
+
+        return $publicSuffix;
     }
 }